Se separo la logica del VerticalDots y se implementó en FeedPost y Post
